requirements.txt
bs4
pymongo
requests
PyExecJS

